xmlrpc
  * Change to version 0.7 on SF

+++ src/xmlrpcpp/XmlRpcSource.cpp
@@ -1,6 +1,21 @@
+// ----------------------------------------------------------------------------
 //
+// flxmlrpc Copyright (c) 2015 by W1HKJ, Dave Freese <iam_w1hkj@w1hkj.com>
+//    
 // XmlRpc++ Copyright (c) 2002-2008 by Chris Morley
 //
+// This file is part of fldigi
+//
+// flxmlrpc is free software; you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation; either version 3 of the License, or
+// (at your option) any later version.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+// ----------------------------------------------------------------------------
+
+#include <config.h>
 
 #include "XmlRpcSource.h"
 #include "XmlRpcUtil.h"
@@ -187,7 +202,7 @@ namespace XmlRpc {
       _ssl = 0;
     }
 #endif
-    if (_fd != -1)
+    if ( (int)_fd != -1)
     {
       XmlRpcUtil::log(2,"XmlRpcSource::close: closing socket %d.", _fd);
       XmlRpcSocket::close(_fd);
